==> 06_Query5_incomes.ipynb <==
import time
from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from sedona.register import SedonaRegistrator

# Dynamic Allocation

spark_conf = {
    "spark.serializer": "org.apache.spark.serializer.KryoSerializer",
    "spark.kryo.registrator": "org.apache.sedona.core.serde.SedonaKryoRegistrator",
    "spark.sql.adaptive.enabled": "true",
    "spark.dynamicAllocation.enabled": "true",
    "spark.shuffle.service.enabled": "true"
}

spark = SparkSession.builder \
    .appName("Query5_Clean_Run") \
    .config(map=spark_conf) \
    .getOrCreate()

SedonaRegistrator.registerAll(spark)
base_path = "s3a://groups-bucket-dblab-905418150721/group36/processed_data"

print("--- Loading Data ---")


blocks = spark.read.parquet(f"{base_path}/census_blocks_geo.parquet").select(
    F.col("features.properties.COMM").getItem(0).alias("COMM"),
    F.col("features.properties.POP20").getItem(0).cast("long").alias("total_pop"),
    F.col("features.properties.ZCTA20").getItem(0).cast("long").cast("string").alias("zcta"),
    F.expr("ST_GeomFromGeoJSON(to_json(features.geometry))").alias("geometry")
)
blocks.createOrReplaceTempView("blocks")

# CRIMES: Sampling 20% 
crimes = spark.read.parquet(f"{base_path}/crime_data_clean.parquet") \
    .filter((F.col("Year") >= 2020)) \
    .sample(False, 0.2, 42) \
    .select("LON", "LAT")
crimes.createOrReplaceTempView("crimes")

# INCOME
spark.read.parquet(f"{base_path}/income_data.parquet").select(
    F.col("Zip Code").cast("long").cast("string").alias("zip_code"),
    F.col("Estimated Median Income").alias("income")
).createOrReplaceTempView("income")

print("--- Executing Join ---")
start_time = time.time()

query = """
SELECT 
    b.COMM,
    AVG(i.income) as avg_income,
    (COUNT(c.LON) * 5) / SUM(b.total_pop) as crime_rate -- x5 λόγω 20% sample
FROM blocks b
LEFT JOIN income i ON b.zcta = i.zip_code
JOIN crimes c ON ST_Contains(b.geometry, ST_Point(CAST(c.LON AS Decimal(24,20)), CAST(c.LAT AS Decimal(24,20))))
WHERE b.total_pop > 0 
GROUP BY b.COMM
HAVING avg_income IS NOT NULL
"""

# Χρησιμοποιούμε limit για να δούμε αν ξεκινάει η ροή
final_df = spark.sql(query)
results = final_df.collect()

print(f"Time: {time.time() - start_time:.2f}s")

if results:
    df = spark.createDataFrame(results)
    print(f"Correlation: {df.stat.corr('avg_income', 'crime_rate'):.5f}")
    df.show(5)
else:
    print("No results.")
